Add files via upload
This is the pdf containing the formal write-up of the project from my Discrete Mathematics class with Professor Crotty(section 007) that I am doing this semester. 
I do not plan on collaborating with a classmate if this is an acceptable final project preposal. 
I am interested in the deep history of cryptography. I have a BA in political science and anyway I can incorporate learning about governmental abuse of power or secrecy in powerful realms, I try to. I think a more formal dive into cryptography will help me with contextualizing some of the more abstract areas of computer science.